backend/requirements.txt
Django>=4.2
djangorestframework
django-filter
psycopg2-binary
django-cors-headers
drf-yasg

